upload changes

+++ project/lib/png/ci/ci_verify_configure.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-set -o errexit -o pipefail -o posix
+set -e
 
 # Copyright (c) 2019-2024 Cosmin Truta.
 #
@@ -8,7 +8,7 @@ set -o errexit -o pipefail -o posix
 #
 # SPDX-License-Identifier: MIT
 
-# shellcheck source=ci/lib/ci.lib.sh
+# shellcheck source="ci/lib/ci.lib.sh"
 source "$(dirname "$0")/lib/ci.lib.sh"
 cd "$CI_TOPLEVEL_DIR"
 
@@ -26,12 +26,10 @@ function ci_init_build {
         [[ -x $(command -v cc) ]] && CI_CC="${CI_CC:-cc}"
     }
     # Ensure that the CI_ variables that cannot be customized reliably are not initialized.
-    [[ ! $CI_CONFIGURE_VARS ]] || {
+    [[ ! $CI_CONFIGURE_VARS ]] ||
         ci_err "unsupported: \$CI_CONFIGURE_VARS='$CI_CONFIGURE_VARS'"
-    }
-    [[ ! $CI_MAKE_VARS ]] || {
+    [[ ! $CI_MAKE_VARS ]] ||
         ci_err "unsupported: \$CI_MAKE_VARS='$CI_MAKE_VARS'"
-    }
 }
 
 function ci_trace_build {
@@ -57,54 +55,31 @@ function ci_trace_build {
     ci_info "environment option: \$CI_LD: '$CI_LD'"
     ci_info "environment option: \$CI_LD_FLAGS: '$CI_LD_FLAGS'"
     ci_info "environment option: \$CI_SANITIZERS: '$CI_SANITIZERS'"
-    ci_info "environment option: \$CI_FORCE: '$CI_FORCE'"
     ci_info "environment option: \$CI_NO_TEST: '$CI_NO_TEST'"
     ci_info "environment option: \$CI_NO_INSTALL: '$CI_NO_INSTALL'"
     ci_info "environment option: \$CI_NO_CLEAN: '$CI_NO_CLEAN'"
     ci_info "executable: \$CI_MAKE: $(command -V "$CI_MAKE")"
-    [[ $CI_CC ]] && {
+    [[ $CI_CC ]] &&
         ci_info "executable: \$CI_CC: $(command -V "$CI_CC")"
-    }
-    [[ $CI_CPP ]] && {
+    [[ $CI_CPP ]] &&
         ci_info "executable: \$CI_CPP: $(command -V "$CI_CPP")"
-    }
-    [[ $CI_AR ]] && {
+    [[ $CI_AR ]] &&
         ci_info "executable: \$CI_AR: $(command -V "$CI_AR")"
-    }
-    [[ $CI_RANLIB ]] && {
+    [[ $CI_RANLIB ]] &&
         ci_info "executable: \$CI_RANLIB: $(command -V "$CI_RANLIB")"
-    }
-    [[ $CI_LD ]] && {
+    [[ $CI_LD ]] &&
         ci_info "executable: \$CI_LD: $(command -V "$CI_LD")"
-    }
     ci_info "## END OF CONFIGURATION ##"
 }
 
 function ci_cleanup_old_build {
-    ci_info "## START OF PRE-BUILD CLEANUP ##"
-    [[ ! -e $CI_BUILD_DIR && ! -e $CI_INSTALL_DIR ]] || {
+    if [[ -e $CI_BUILD_DIR || -e $CI_INSTALL_DIR ]]
+    then
+        ci_info "## START OF PRE-BUILD CLEANUP ##"
         ci_spawn rm -fr "$CI_BUILD_DIR"
         ci_spawn rm -fr "$CI_INSTALL_DIR"
-    }
-    [[ ! -e "$CI_SRC_DIR/config.status" ]] || {
-        ci_warn "unexpected build configuration file: '$CI_SRC_DIR/config.status'"
-        if ci_expr $((CI_FORCE))
-        then
-            # Delete the old config and (possibly) the old build.
-            ci_info "note: forcing an in-tree build cleanup"
-            if [[ -f $CI_SRC_DIR/Makefile ]]
-            then
-                ci_spawn make -C "$CI_SRC_DIR" distclean
-            else
-                ci_spawn rm -fr "$CI_SRC_DIR"/config.{log,status}
-            fi
-        else
-            # Alert the user, but do not delete their files.
-            ci_warn "the configure script might fail"
-            ci_info "hint: consider using the option \$CI_FORCE=1"
-        fi
-    }
-    ci_info "## END OF PRE-BUILD CLEANUP ##"
+        ci_info "## END OF PRE-BUILD CLEANUP ##"
+    fi
 }
 
 function ci_build {
@@ -119,14 +94,10 @@ function ci_build {
     [[ $CI_LD ]] && ci_spawn export LD="$CI_LD"
     [[ $CI_LD_FLAGS ]] && ci_spawn export LDFLAGS="$CI_LD_FLAGS"
     [[ $CI_SANITIZERS ]] && {
-        ci_spawn export CFLAGS="${CFLAGS:-"-O2"} -fsanitize=$CI_SANITIZERS"
-        ci_spawn export LDFLAGS="${LDFLAGS}${LDFLAGS:+" "}-fsanitize=$CI_SANITIZERS"
+        ci_spawn export CFLAGS="-fsanitize=$CI_SANITIZERS ${CFLAGS:-"-O2"}"
+        ci_spawn export LDFLAGS="-fsanitize=$CI_SANITIZERS $LDFLAGS"
     }
-    # Spawn "autogen.sh" if the configure script is not available.
-    [[ -x "$CI_SRC_DIR/configure" ]] || {
-        ci_spawn "$CI_SRC_DIR/autogen.sh" --maintainer
-    }
-    # And... build!
+    # Build!
     ci_spawn mkdir -p "$CI_BUILD_DIR"
     ci_spawn cd "$CI_BUILD_DIR"
     # Spawn "configure".
@@ -150,9 +121,8 @@ function ci_build {
 }
 
 function usage {
-    echo "usage: $CI_SCRIPT_NAME [<options>]"
-    echo "options: -?|-h|--help"
-    exit "${@:-0}"
+    echo "usage: $CI_SCRIPT_NAME"
+    exit 0
 }
 
 function main {
@@ -160,7 +130,7 @@ function main {
     while getopts ":" opt
     do
         # This ain't a while-loop. It only pretends to be.
-        [[ $1 == -[?h]* || $1 == --help || $1 == --help=* ]] && usage 0
+        [[ $1 == -[?h]* || $1 == --help ]] && usage
         ci_err "unknown option: '$1'"
     done
     shift $((OPTIND - 1))
@@ -168,9 +138,8 @@ function main {
     ci_init_build
     ci_trace_build
     [[ $# -eq 0 ]] || {
-        echo >&2 "error: unexpected argument: '$1'"
-        echo >&2 "note: this program accepts environment options only"
-        usage 2
+        ci_info "note: this program accepts environment options only"
+        ci_err "unexpected argument: '$1'"
     }
     ci_cleanup_old_build
     ci_build
